% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_pred_planned.R
\name{diff_pred_planned}
\alias{diff_pred_planned}
\title{Calculating difference between planned and predicted sales}
\usage{
diff_pred_planned(.pred, .pred_date = "date", .planned,
  .planned_date = "date", .level = "total", .key = "retail_sales_m3",
  .horizon = 1, .total_horizons = 3, .total_sales = NULL,
  .date_start = NULL, .date_end = NULL)
}
\arguments{
\item{.pred}{List which contains a data frame for each level and horizon regarding the predictions.}

\item{.pred_date}{Date column of .pred input}

\item{.planned}{List which contains a data frame for each level and horizon regarding the planned sales.}

\item{.planned_date}{Date column of .planned input}

\item{.level}{Level of interest. Input can be a single string or a character vector.}

\item{.key}{Key variable which is used for calculating the difference.}

\item{.horizon}{The number of time horizons which are included in .pred.}

\item{.total_horizons}{The number of total horizons which are included in .pred.}

\item{.total_sales}{Parameter without any usefull functionality. But there can be something in the future.}

\item{.date_start}{Parameter without any usefull functionality. But there can be something in the future.}

\item{.date_end}{Parameter without any usefull functionality. But there can be something in the future.}
}
\description{
This function calculates the difference between planned and predicted sales.
}
\details{
If someone is interested for a detailed documentation: Go Home!
}
